Glitch

+++ server.js
@@ -5,6 +5,7 @@ const http = require('http');
 const ms = require('ms');
 const express = require('express');
 const randomPuppy = require('random-puppy');
+
 const app = express();
 app.get("/", (request, response) => {
  
@@ -33,7 +34,6 @@ const config = require("./config.json");
 //----------
 
 
-
 //----------
 
 
@@ -46,9 +46,9 @@ const config = require("./config.json");
    ];
    setInterval(function() {
      
-     let status = statuses[Math.floor(Math.random()*statuses.length)];
+     var botstatus = statuses[Math.floor(Math.random()*statuses.length)];
      
-     client.user.setPresence({ game : { name: status }, status: "online"})
+     client.user.setPresence({ game : { name: botstatus }, status: "online"})
    }, 60000)
  });
 
@@ -68,7 +68,9 @@ const config = require("./config.json");
 client.on("guildCreate", guild => {
   // This event triggers when the bot joins a guild.
   console.log(`Nouveau discord: ${guild.name} (id: ${guild.id}). Ce discord a ${guild.memberCount} membres!`);
-  client.user.setActivity(`s!help | Sur ${client.guilds.size} serveurs`);
+client.user.setActivity(`s!help | Sur ${client.guilds.size} serveurs`);
+     
+
    var found = false;
   guild.channels.forEach(function(channel, id) {
       if(found == true || channel.type != "text") {
@@ -77,7 +79,7 @@ client.on("guildCreate", guild => {
       if(guild.me.permissionsIn(channel).has("SEND_MESSAGES") && guild.me.permissionsIn(channel).has("VIEW_CHANNEL")) {
         found = true;
         return channel.send("Hey, je suis Saturn, Merci de m'avoir invité dans le serveur! Voir mes commandes est facile, faites s!help.")
-        console.log("Server joined")
+        
       }
   })
 });
@@ -86,7 +88,7 @@ client.on("guildCreate", guild => {
 client.on("guildDelete", guild => {
   // this event triggers when the bot is removed from a guild.
   console.log(`J'ai été retiré de: ${guild.name} et l'ID est :(id: ${guild.id})`);
-  client.user.setActivity(`s!help | Sur ${client.guilds.size} serveurs`);
+client.user.setActivity(`s!help | Sur ${client.guilds.size} serveurs`);
 });
 client.on("guildMemberAdd", (member) => { // Check out previous chapter for information about this event
 let guild = member.guild; 
@@ -101,6 +103,7 @@ if(guild.systemChannel){
 	.setTimestamp() // Sets a timestamp at the end of the embed
 	);
 }
+    console.log(`${memberTag} a rejoins le discord ${guild.name}. Ce discord a ${guild.memberCount} membres!`);
 });
 client.on("guildMemberRemove", (member) => { // Check out previous chapter for information about this event
 let guild = member.guild; 
@@ -115,6 +118,7 @@ if(guild.systemChannel){
 	.setTimestamp() // Sets a timestamp at the end of the embed
 	);
 }
+   console.log(`${memberTag} a quitté le discord ${guild.name}. Ce discord a ${guild.memberCount} membres!`);
 });
 
 
@@ -143,10 +147,12 @@ client.on("message", async message => {
   // Let's go with a few common example commands! Feel free to delete or change those.
   
   if(command === "ping") {
+
     // Calculates ping between sending a message and editing it, giving a nice round-trip latency.
     // The second ping is an average latency between the bot and the websocket server (one-way, not round-trip)
     const m = await message.channel.send("Attends une seconde...");
     m.edit(`Pong! :ping_pong: La latence est de ${m.createdTimestamp - message.createdTimestamp}ms. La latence de l'API est de ${Math.round(client.ping)}ms`);
+       console.log(`${message.author.tag} a utilisé la commande Ping dans le discord ${message.guild.name}.`);
   }
   if(command === "meme") {
             const subReddits = ["dankmeme", "meme", "me_irl"];
@@ -161,12 +167,14 @@ client.on("message", async message => {
             .setFooter(`Si l'image n'apparait pas immédiatement, veuillez attendre quelque secondes pour que l'image charge :)`)
 
         message.channel.send(memeembed);
+           console.log(`${message.author.tag} a utilisé la commande Meme dans le discord ${message.guild.name}.`);
   }
   if(command === "annonce") {
      const annonceMessage = args.join(" ");
  if(message.author.id !== config.ownerID) return message.reply("Desolé, mais tu n'a pas accès à cette commande!");                                                                                       
      message.delete().catch(O_o=>{}); 
-    client.channels.get("639994872689983518").send(annonceMessage)     
+    client.channels.get("650422877258252298").send(annonceMessage)    
+   console.log(`${message.author.tag} a utilisé la commande Annonce dans le discord ${message.guild.name}.`);
 };
                                                                                       
 
@@ -187,6 +195,7 @@ client.on("message", async message => {
                             message.react("❌")
             message.channel.send(`${err}`, {code:"js"})
                     }
+           console.log(`${message.author.tag} a utilisé la commande Eval dans le discord ${message.guild.name}.`);
             }
  
  
@@ -205,9 +214,10 @@ client.on("message", async message => {
    .addField("Erreur / bug:", `${args.join(' ')}`, true)
      .addField("Type d'erreur", "Erreur sur le Discord.")
     .setFooter("Avis de report - Saturn")
-  client.channels.get("639997214403592192").send(reportEmbed)
+  client.channels.get("650426196496089148").send(reportEmbed)
    
     message.reply("Merci de ton soutien! Je vais regarder ca de plus près.")
+         console.log(`${message.author.tag} a utilisé la commande ReportError dans le discord ${message.guild.name}. \nErreur reporté: ${args.join(' ')}`);
   }
   
 	  if (command === "restart") {
@@ -215,6 +225,7 @@ client.on("message", async message => {
      if(message.author.id !== config.ownerID) return message.reply("Desolé, mais tu n'a pas accès à cette commande!");   
             message.react("👍");
             console.log("Redémarrage du bot en cours...");
+      console.log(`${message.author.tag} a utilisé la commande Restart dans le discord ${message.guild.name}.`);
       await message.channel.send("Redémarrage du bot maintenant!")
       client.destroy();
 
@@ -227,20 +238,22 @@ client.on("message", async message => {
      if(message.author.id !== config.ownerID) return message.reply("Desolé, mais tu n'a pas accès à cette commande!");   
             message.react("👍");
             console.log("Fermeture du bot en cours...");
+       console.log(`${message.author.tag} a utilisé la commande Shutdown dans le discord ${message.guild.name}.`);
       await message.channel.send("Fermeture du bot maintenant!")
             client.destroy();
 
           
         }
   if (command === "setstatus") {
-     let statusArgs = args.join("  ");;
+     let statusArgs = args.join("  ");
      if(message.author.id !== config.ownerID) return message.reply("Desolé, mais tu n'a pas accès à cette commande!");   
             message.react("👍");
             console.log("Changement du status en cours...");
       await message.channel.send("Changement du status en cours...")
          client.user.setActivity(statusArgs);
     message.channel.send("Status changé!")
-
+ console.log(`Changement du status a: ${statusArgs}`);
+     console.log(`${message.author.tag} a utilisé la commande SetStatus dans le discord ${message.guild.name}.`);
           
         }
                                     
@@ -269,6 +282,7 @@ client.on("message", async message => {
 })
 	   
     message.reply("Merci de ton soutien! Je vais regarder ton erreur.")
+      console.log(`${message.author.tag} a utilisé la commande ErrorMP dans le discord ${message.guild.name}.`);
     
 };
    
@@ -281,17 +295,16 @@ client.on("message", async message => {
     // To get the "message" itself we join the `args` back into a string with spaces: 
     
     const sayMessage = args.join(" ");
-    
+    if(!sayMessage) return message.reply("Tu n'as pas écrit de texte a répéter!")
     // Then we delete the command message (sneaky, right?). The catch just ignores the error with a cute smiley thing.
     message.delete().catch(O_o=>{}); 
     // And we get the bot to say the thing:
     message.channel.send(sayMessage);
  
-   
+   console.log(`${message.author.tag} a utilisé la commande Say dans le discord ${message.guild.name}. L'utilisateur a dit ${sayMessage}`);
       
     }
   
-
 
 
 
@@ -300,20 +313,22 @@ client.on("message", async message => {
 
     if (command === "8ball") {
      
-      var sayings = ["Oui", "Non", "Peut-etre", "Peut-etre pas", "Certainement", "Certainement pas", "Bien sur!", "Pas-sur...",]
+      var sayings = ["Oui", "Non", "Peut-etre", "Peut-etre pas", "Certainement", "Certainement pas", "Bien sur!", "Pas sur...",]
   const ballText = args.join(" ");
 			var result = Math.floor((Math.random() * sayings.length) + 0);
       if(!ballText)
       return message.reply("Veuillez demander une question pour utiliser 8ball :8ball: ")
 			const embed = new Discord.RichEmbed()
      .setColor("RANDOM")
-    .setTitle(":8ball:ball")
+    .setTitle("8ball")
     .addField(ballText, sayings[result])
     message.channel.send({embed:embed})
+      console.log(`${message.author.tag} a utilisé la commande 8ball dans le discord ${message.guild.name}.`);
     }
 
   
 if(command === "dm") {
+   if(message.author.id !== config.ownerID) return message.reply("Desolé, mais tu n'a pas accès à cette commande!");   
    let dmMember = message.mentions.members.first() || message.guild.members.get(args[0]);
     if(!dmMember)
       return message.reply("Mentionnez une personne dans le serveur");
@@ -326,8 +341,9 @@ if(command === "dm") {
    
     
     // Now, time for a swift kick in the nuts!
-    await dmMember.send(dmArgs)
+    await dmMember.send(`${message.author.tag} vous a envoyé ce message: ` + dmArgs)
     message.reply(`Message envoyé a ${dmMember.user.tag}.`)
+  console.log(`${message.author.tag} a utilisé la commande DM dans le discord ${message.guild.name}.`);
 };
 
   
@@ -368,6 +384,7 @@ if(command === "dm") {
   
     .setFooter("Avis de kickement - Saturn")
     message.channel.send(kick_embed)
+    console.log(`${message.author.tag} a utilisé la commande Kick dans le discord ${message.guild.name}.`);
     }
   
 
@@ -417,6 +434,7 @@ if(command === 'mute') {
         message.delete()
         tomute.send(mutedembed)
         message.channel.send(`${tomute.user.username} a été muté pendant ${ms(ms(mutetime))}!`)
+        console.log(`${message.author.tag} a utilisé la commande Mute dans le discord ${message.guild.name}.`);
       })
     setTimeout(function(){
     tomute.removeRole(muterole.id);
@@ -453,6 +471,7 @@ if(command === 'mute') {
   .addField("Serveurs:",`${client.guilds.array().sort()}`)
   .setFooter("Serverlist - Saturn")
       message.channel.send(serverlist_embed)
+   console.log(`${message.author.tag} a utilisé la commande Serverlist dans le discord ${message.guild.name}.`);
 };
  
   
@@ -517,6 +536,7 @@ if(command === 'mute') {
     message.channel.send(warnSuccessfulEmbed); // Sends the warn successful embed
 
     message.delete(); // Deletes the command
+    console.log(`${message.author.tag} a utiliser la commande Warn dans le discord ${message.guild.name}.`);
 }
   
   if(command === "ban") {
@@ -545,6 +565,7 @@ if(command === 'mute') {
   
     .setFooter("Avis de bannisement - Saturn")
     message.channel.send(ban_embed)
+    console.log(`${message.author.tag} a utiliser la commande Ban dans le discord ${message.guild.name}.`);
     }
   
   
@@ -563,7 +584,7 @@ if(command === "help"){
   .setFooter("Help - Saturn")
   message.reply("Regarde tes MPs! :eyes:")
   message.author.send(help_embed)
-  console.log("Un utilisateur a demandé de l'aide")
+ console.log(`${message.author.tag} a utiliser la commande Help dans le discord ${message.guild.name}.`);
 }
   
   if(command === "debug"){
@@ -579,6 +600,7 @@ if(command === "help"){
   .addField("Créateur:", "LunatiikXD")
   .setFooter("Debug - Saturn")
   message.channel.send(debug_embed)
+    console.log(`${message.author.tag} a utiliser la commande Debug dans le discord ${message.guild.name}.`);
 }
   
   if(command === "invite"){
@@ -588,6 +610,7 @@ if(command === "help"){
   .addField("Invitez le bot sur votre serveur!", "https://discordapp.com/oauth2/authorize?client_id=639248385794113563&scope=bot&permissions=805314622")
   .setFooter("Invite - Saturn")
   message.channel.send(inv_embed);
+    console.log(`${message.author.tag} a utiliser la commande Invite dans le discord ${message.guild.name}.`);
   }
   
  
@@ -597,14 +620,15 @@ if(command === "help"){
     .setDescription("**Info sur moi**")
     .addField("Nom:", "Saturn", true)
     .addField("Créateur:", "LunatiikXD", true)
-    .addField("Version de discord.js:", `${require('discord.js').version}`)
+    .addField("Version de discord.js:", `${require('discord.js').version}`, false)
     .addField("Version du bot:","1.1.0", true)
     .addField(`Serveurs:`, `${client.guilds.size} serveur(s)`, false) 
     .addField(`Utilisateurs:`, `${client.users.size} personnes utilisent Saturn`, false)
-    .addField("Serveur Discord:", "https://discord.gg/qCxgwp5")
+    .addField("Discord:", "https://discord.gg/BZJ6SBk")
     .setThumbnail(client.user.avatarURL)
     .setFooter("Info - Saturn")
     message.channel.send(inf_embed)
+       console.log(`${message.author.tag} a utiliser la commande Info dans le discord ${message.guild.name}.`);
     
   //message.member.send(`:eyes:Info sur moi:eyes:\nNom: Mars\nCreateur: LunatiikXD\nCreer avec discord.js 11.4.0\nJe suis sur ${client.guilds.size} serveurs !`)
  
@@ -623,6 +647,7 @@ if(command === "userinfo") {
     .addField("Tu a rejoins ce discord le", message.member.joinedAt, true)
     .setFooter("Userinfo - Saturn")
     message.channel.send(userEmbed)
+  console.log(`${message.author.tag} a utiliser la commande Userinfo dans le discord ${message.guild.name}.`);
   }
   
   if(command === "serverinfo") {
@@ -634,10 +659,11 @@ if(command === "userinfo") {
     .addField("Crée le", message.guild.createdAt, true)
     .addField("Tu a rejoins le", message.member.joinedAt, true)
     .addField("Membres sur le Discord", message.guild.memberCount, false)
-    .addField("Propriétaire du Discord:", message.guild.owner.user.tag, false)
+    .addField("Propriétaire du Discord:", message.guild.owner.user.tag, true)
     
     .setFooter("Serverinfo - Saturn")
     message.channel.send(servEmbed)
+    console.log(`${message.author.tag} a utiliser la commande Serverinfo dans le discord ${message.guild.name}.`);
   }
   
   if(command === "avatar") {
@@ -647,6 +673,7 @@ if(command === "userinfo") {
 	  .setImage(message.author.avatarURL)
 	  .setFooter("Avatar - Saturn")
 	  message.channel.send(avatar_embed);
+    console.log(`${message.author.tag} a utiliser la commande Avatar dans le discord ${message.guild.name}.`);
   }
 
 
@@ -675,6 +702,7 @@ if(command === "userinfo") {
     setTimeout(function(){ 
       message.channel.bulkDelete(1)
 }, 2000);
+    console.log(`${message.author.tag} a utiliser la commande Clear dans le discord ${message.guild.name}.`);
